+++ examples/esp32/apps/memfault_demo_app/main/cmd_wifi.c
@@ -97,7 +97,11 @@ bool wifi_join(const char *ssid, const char *pass) {
 
     ESP_ERROR_CHECK(esp_netif_init());
 
-    ESP_ERROR_CHECK_WITHOUT_ABORT(esp_event_loop_create_default());
+    // Try to create default event loop. If it's already created, it returns
+    // ESP_ERR_INVALID_STATE, equivalent to ESP_OK for our purposes.
+    esp_err_t err = esp_event_loop_create_default();
+    ESP_ERROR_CHECK_WITHOUT_ABORT(err == ESP_ERR_INVALID_STATE ? ESP_OK : err);
+
     esp_netif_create_default_wifi_sta();
 
     wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();